feat: move interface
