+++ contracts/database/gorm/wire_interface.go
@@ -6,7 +6,7 @@ import (
 	gormio "gorm.io/gorm"
 
 	"github.com/goravel/framework/contracts/config"
-	"github.com/goravel/framework/database/gorm"
+	"github.com/goravel/framework/contracts/database/orm"
 )
 
 type Gorm interface {
@@ -14,20 +14,6 @@ type Gorm interface {
 }
 
 type Initialize interface {
-	InitializeGorm(config config.Config, connection string) *gorm.GormImpl
-	InitializeQuery(ctx context.Context, config config.Config, connection string) (*gorm.QueryImpl, error)
-}
-
-type InitializeImpl struct{}
-
-func NewInitializeImpl() *InitializeImpl {
-	return &InitializeImpl{}
-}
-
-func (receive *InitializeImpl) InitializeGorm(config config.Config, connection string) *gorm.GormImpl {
-	return gorm.InitializeGorm(config, connection)
-}
-
-func (receive *InitializeImpl) InitializeQuery(ctx context.Context, config config.Config, connection string) (*gorm.QueryImpl, error) {
-	return gorm.InitializeQuery(ctx, config, connection)
+	InitializeGorm(config config.Config, connection string) Gorm
+	InitializeQuery(ctx context.Context, config config.Config, connection string) (orm.Query, error)
 }
